producer/requirements.txt
requests==2.31.0
happybase==1.2.0
thrift==0.16.0
kafka-python==2.0.2